requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
python-dotenv>=1.1.0,<2.0
requests>=2.32,<3.0
django-countries>=7.5.1,<8.0
gunicorn>=21.2,<22.0
